Create a separate package for security configs
